feat: add register agent translations for representing name pages
