+++ packages/forms-web-app/src/pages/projects/register/agent/organisation-name/view.njk
@@ -5,7 +5,7 @@
 {% from "govuk/components/input/macro.njk" import govukInput %}
 
 {% set pageHeading = t('register.organisationName.pageHeading')%}
-{% set whoRegisteringFor = t('register.whoRegisteringFor.organisation') %}
+{% set whoRegisteringFor = t('register.whoRegisteringFor.agent') %}
 {% set commonPageTitle = t('register.index.pageTitle') %}
 {% set title = pageHeading + ' - ' + whoRegisteringFor + ' - ' + commonPageTitle %}
