+++ packages/forms-web-app/src/pages/projects/register/agent/representing-organisation-name/view.njk
@@ -4,11 +4,14 @@
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 {% from "govuk/components/input/macro.njk" import govukInput %}
 
-{% set title = "What is the full name of the organisation or charity that you are representing? - Registering on behalf of someone else - Register to have your say about a national infrastructure project - National Infrastructure Planning" %}
+{% set pageHeading = t('register.nameOrganisationRepresenting.pageHeading')%}
+{% set whoRegisteringFor = t('register.whoRegisteringFor.agent') %}
+{% set commonPageTitle = t('register.index.pageTitle') %}
+{% set title = pageHeading + ' - ' + whoRegisteringFor + ' - ' + commonPageTitle %}
 
 {% block pageTitle %}
   {% if errors %}
-    {% set title = "Error: " + title %}
+    {% set title = t('common.error') + ": " + title %}
   {% endif %}
   {{ title }}
 {% endblock %}
@@ -18,7 +21,7 @@
     <div class="govuk-grid-column-two-thirds">
       {% if errorSummary %}
         {{ govukErrorSummary({
-          titleText: "There is a problem",
+          titleText: t('common.thereIsAProblem'),
           errorList: errorSummary
         }) }}
       {% endif %}
@@ -27,7 +30,7 @@
         <div class="govuk-form-group {{ 'govuk-form-group--error' if errors }}">
           {{ govukInput({
             label: {
-              text: "What is the full name of the organisation or charity that you are representing?",
+              text: pageHeading,
               classes: "govuk-label--l",
               isPageHeading: true
             },
@@ -45,7 +48,7 @@
         <input type="hidden" name="representing" value={{ representing }}>
 
         {{ govukButton({
-          text: "Continue",
+          text: t('common.continue'),
           type: "Submit",
           attributes: { "data-cy":"button-save-and-continue"}
         }) }}
