+++ packages/forms-web-app/src/pages/projects/register/agent/representing-who/view.njk
@@ -5,7 +5,7 @@
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 
 {% set pageHeading = t('register.whoRepresenting.pageHeading') %}
-{% set whoRegisteringFor = t('register.whoRegisteringFor.organisation') %}
+{% set whoRegisteringFor = t('register.whoRegisteringFor.agent') %}
 {% set commonPageTitle = t('register.index.pageTitle') %}
 {% set title = pageHeading + ' - ' + whoRegisteringFor + ' - ' + commonPageTitle %}
